+++ python/ray/data/_internal/execution/operators/randomize_block_order_operator.py
@@ -6,14 +6,17 @@ from ray.data._internal.execution.interfaces import PhysicalOperator, RefBundle
 from ray.data._internal.execution.operators.base_physical_operator import (
     OneToOneOperator,
 )
+from ray.data._internal.stats import StatsDict
 from ray.data.block import BlockMetadata
 
-from ray.data._internal.stats import StatsDict
-
+# When NEW_DATA_TO_SHUFFLE_RATIO * window_size new blocks have been added,
+# we do a shuffle.
 NEW_DATA_TO_SHUFFLE_RATIO = 0.3
 
 
 class RandomizeBlockOrderOperator(OneToOneOperator):
+    """Physical operator for randomize_block_order."""
+
     def __init__(
         self,
         input_op: PhysicalOperator,
@@ -26,9 +29,10 @@ class RandomizeBlockOrderOperator(OneToOneOperator):
         super().__init__(self._name, input_op)
         self._window_size = window_size
         self._random = Random(seed)
-        self._out_buffer: List[RefBundle] = []
+        self._out_buffer: Deque[RefBundle] = deque()
         self._output_metadata: List[BlockMetadata] = []
         self._num_outputs_total = input_op.num_outputs_total()
+        # Number of new blocks added since the last shuffle.
         self._num_new_blocks_since_last_shuffle = 0
 
     def add_input(self, refs: RefBundle, input_index: int) -> None:
@@ -39,7 +43,9 @@ class RandomizeBlockOrderOperator(OneToOneOperator):
 
     def all_inputs_done(self) -> None:
         super().all_inputs_done()
-        self._random.shuffle(self._out_buffer)
+        if self._window_size is None:
+            # If window_size is None, do a global shuffle when all inputs are received.
+            self._random.shuffle(self._out_buffer)
 
     def has_next(self) -> bool:
         if self._window_size is None:
@@ -62,8 +68,7 @@ class RandomizeBlockOrderOperator(OneToOneOperator):
 
     def get_next(self) -> RefBundle:
         self._maybe_shuffle()
-        index = self._random.randint(0, len(self._out_buffer) - 1)
-        res = self._out_buffer.pop(index)
+        res = self._out_buffer.popleft()
         self._output_metadata.extend([meta for _, meta in res.blocks])
         return res
